backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
httpx==0.27.0
asyncpg==0.29.0
sqlalchemy[asyncio]==2.0.30
pydantic==2.7.0
pydantic-settings==2.3.0
python-dotenv==1.0.1
trafilatura==1.10.0
feedparser==6.0.11
youtube-transcript-api==0.6.2
google-cloud-texttospeech==2.16.3
dashscope
